[#4] Split Assignee class
